Fix routing problem on the ExpensesIncomeScreen, fix logout logic
