+++ Budgety-Native-FE/src/data/constants.ts
@@ -1,0 +1,9 @@
+export const CATEGORY_FORM_TYPES = {
+  ADD: 'ADD',
+  EDIT: 'EDIT'
+} as const;
+
+export const CONTEXT = {
+  EXPENSES: 'EXPENSES',
+  INCOME: 'INCOME'
+} as const;
