+++ Budgety-Native-FE/src/routes/DrawerNavigator/DrawerNavigator.tsx
@@ -1,8 +1,8 @@
 import { COLORS } from '../../styles/Colors';
 import { createDrawerNavigator } from '@react-navigation/drawer';
 import CustomDrawerContent from '../CustomDrawerContent/CustomDrawerContent';
+import ExpensesIncomeScreen from '../../screens/ExpensesIncomeScreen/ExpensesIncomeScreen';
 import ExchangeScreen from '../../screens/ExchangeScreen/ExchangeScreen';
-import ExpensesIncomeStack from '../ExpensesIncomeStack/ExpensesIncomeStack';
 import HistoryScreen from '../../screens/HistoryScreen/HistoryScreen';
 import HomeScreen from '../../screens/HomeScreen/HomeScreen';
 import SettingsScreen from '../../screens/SettingsScreen/SettingsScreen';
@@ -25,7 +25,7 @@ const DrawerNavigator: React.FC = () => (
     }}
   >
     <Drawer.Screen component={HomeScreen} name="Home" />
-    <Drawer.Screen component={ExpensesIncomeStack} name="ExpensesIncome" />
+    <Drawer.Screen component={ExpensesIncomeScreen} name="ExpensesIncome" />
     <Drawer.Screen component={ExchangeScreen} name="Exchange" />
     <Drawer.Screen component={HistoryScreen} name="History" />
     <Drawer.Screen component={SettingsScreen} name="Settings" />
