+++ Budgety-Native-FE/src/slices/settingsSlice.ts
@@ -1,6 +1,6 @@
 import { createSlice, PayloadAction } from '@reduxjs/toolkit';
 
-interface SettingsState {
+export interface SettingsState {
   currency: string;
   language: string;
   username: string;
