+++ Budgety-Native-FE/src/store/store.ts
@@ -1,12 +1,30 @@
-import { configureStore } from '@reduxjs/toolkit';
-import settingsSlice from '../slices/settingsSlice';
-import userSlice from '../slices/userSlice';
+import { AnyAction, configureStore } from '@reduxjs/toolkit';
+import expenseIncomeSlice, {
+  IExpenseIncomeSlice
+} from '../slices/expenseIncomeSlice';
+import settingsSlice, { SettingsState } from '../slices/settingsSlice';
+import userSlice, { UserSlice, logout } from '../slices/userSlice';
+
+interface IState {
+  expensesIncome: IExpenseIncomeSlice;
+  settings: SettingsState;
+  user: UserSlice;
+}
+
+const rootReducer = (state: IState | undefined, action: AnyAction) => {
+  if (action.type === logout.type) {
+    state = undefined;
+  }
+  console.log('state', state);
+  return {
+    expensesIncome: expenseIncomeSlice(state?.expensesIncome, action),
+    settings: settingsSlice(state?.settings, action),
+    user: userSlice(state?.user, action)
+  };
+};
 
 export const store = configureStore({
-  reducer: {
-    settings: settingsSlice,
-    user: userSlice
-  }
+  reducer: rootReducer
 });
 
 export type AppDispatch = typeof store.dispatch;
